Fix a corner case in the ATX header parser
The parser would previously crash in some cases of empty ATX headers with
trailing hashes. This case isn't captured by the CommonMark spec.

The bad cases look like this:

```
 ###    ####
```

(note the leading space here is not part of the bug - it's just so `git commit`
doesn't throw these lines away as comments)

In these cases the line parser advances `first_nonspace` all the way to the
second set of hashes, while the `chop_trailing_spaces` routine for stripping
the trailing hashes also trims leading whitespace, leading to a buffer
overrun as `add_text_to_container` thinks the content of the line
is only the leading "###" but that the text begins at the start
of the start of "####".

This hasn't been detected previously because the spec only contains
a test case for

```
 ### ###
```

and standard CommonMark always eats exactly one space after the initial ATX
hashes, leaving the line parser's `offset` equal to `first_nonspace`, and
correctly adding empty text to the header container.

The spec might want an additional test case containing multiple spaces in an
empty bookended ATX header.