Hoists header -> anchor code to where it can be used and tested independently
Converting headers to anchors is useful even outside the direct context
of converting markdown to html.  This commit makes the pre-existing code
that does such conversions available as a standalone struct, so that
people who want to convert a header to an anchor don't have to re-implement
the algorithm themselves.