parser: also reset these fields
Likewise from https://github.com/github/cmark-gfm/pull/95; unfortunately
the quadratic behaviour is still present.

Reversing the same changes in upstream cmark-gfm today does *not* cause
the behaviour to regress, leading me to believe there are other changes
made upstream since that stop this from becoming an issue to begin with.

These changes are still correct (and map to upstream), so I'll leave
them in, but the issue remains.