scanners: don't overrun
NULs are not allowed in the input stream (which re2c handily enforces
for us), so we can return it at EOF and it Just Works(tm).

I'm not sure if adding this length check is going to be really expensive
or not, though.  Upstream takes care of this by caching the byte at the
scan boundary, writing a NUL there before scanning, then restoring it
after.  Memory-unsafety aside, we may not even have the literal memory
available for us in all cases -- we can't rely on there necessarily
being (room for) one at the end of the string in all calls to the lexer
as it is -- so it's a bit out of reach.