Require domains for extended URL autolinks to have periods
This commit covers an issue where the number of periods in a domain was
not validated for an extended URL autolink. This was clearly
intentional, as there was a dedicated parameter to loosen the check, and
tests for it. However, it means that bare schemes such as "http://" are
autolinked.