Be stricter about non-relaxed autolinks
Require that valid domains contain at least one period if and only if
we're using relaxed_autolinks. Previously, this was never enforced for
extended URL autolinks, meaning "http://" was a valid autolink. For
extended www autolinks, this was always enforced, but would always pass
as the domain validation included the initial "www.". We now skip that
for domain validation.