redo cm with str internally.
We could do much better than this; I think we should be able to
periodically flush the output buffer.

Other than reading the very last byte in a few places (LF check in
output, digit check in outc), the only read of self.output is when
width>0, in which case that site becomes the place we can flush.