append onto a consistent target.
This stops us from creating a larger intermediate value when there's a
Text node on either side of an Escaped Text node; we append the Escaped
Text to the left-hand Text, and then the right Text on the left-hand
Text.

Previously, we were appending the Escaped Text to the right-hand Text,
and then appending _that_ to the left-hand Text.